fix(svg): update svg export check imports
