+++ src/gradient/Gradient.ts
@@ -3,8 +3,9 @@ import { parseTransformAttribute } from '../parser/parseTransformAttribute';
 import type { FabricObject } from '../shapes/Object/FabricObject';
 import type { TMat2D } from '../typedefs';
 import { uid } from '../util/internals/uid';
+import { isSafeSvgStyleValue } from '../util/internals/svgExportCheck';
 import { pick } from '../util/misc/pick';
-import { isSafeSvgStyleValue, matrixToSVG } from '../util/misc/svgExport';
+import { matrixToSVG } from '../util/misc/svgExport';
 import { linearDefaultCoords, radialDefaultCoords } from './constants';
 import { parseColorStops } from './parser/parseColorStops';
 import { parseCoords } from './parser/parseCoords';
