+++ src/util/index.ts
@@ -62,8 +62,8 @@ export {
   getSafeSvgStyleNumber,
   getSafeSvgStyleToken,
   isSafeSvgStyleValue,
-  matrixToSVG,
-} from './misc/svgExport';
+} from './internals/svgExportCheck';
+export { matrixToSVG } from './misc/svgExport';
 export { groupSVGElements } from './misc/groupSVGElements';
 export { findScaleToFit, findScaleToCover } from './misc/findScaleTo';
 export { capValue } from './misc/capValue';
